Time: 0 ms (100.00%) | Memory: 8.9 MB (22.27%) - LeetSync

+++ 9-palindrome-number/palindrome-number.c
@@ -1,6 +1,6 @@
 bool isPalindrome(int x) {
     int digit =0;
-    int long long  reverse =0;
+    int long   reverse =0;
     int original = x;
 
     if(x<0){
